+++ src/main/java/com/falsepattern/zigbrains/zon/highlight/ZonSyntaxHighlighterFactory.java
@@ -1,0 +1,15 @@
+package com.falsepattern.zigbrains.zon.highlight;
+
+import com.intellij.openapi.fileTypes.SyntaxHighlighter;
+import com.intellij.openapi.fileTypes.SyntaxHighlighterFactory;
+import com.intellij.openapi.project.Project;
+import com.intellij.openapi.vfs.VirtualFile;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
+
+public class ZonSyntaxHighlighterFactory extends SyntaxHighlighterFactory {
+    @Override
+    public @NotNull SyntaxHighlighter getSyntaxHighlighter(@Nullable Project project, @Nullable VirtualFile virtualFile) {
+        return new ZonSyntaxHighlighter();
+    }
+}